Documentation additions, added NetworkCallsInternal.
